entra: organize authentication types
The Entra-renamed authentication implementation was still living
directly under src/Core/Authentication and sharing the general
GitCredentialManager.Authentication namespace with unrelated
authentication helpers (basic auth, NTLM, OAuth, etc). Give it a
dedicated home to make the boundary between the Entra ID
implementation and the rest of the authentication code obvious.

Move EntraAuthentication.cs and WorkloadFederationOptions.cs into
src/Core/Authentication/Entra/, and the corresponding tests into
src/Core.Tests/Authentication/Entra/, under the new
GitCredentialManager.Authentication.Entra namespace. Extract
IEntraAuthentication/IEntraAuthenticationResult and
ServicePrincipalIdentity out of the EntraAuthentication.cs monolith
into their own files, since they are separate public contracts rather
than implementation detail of the EntraAuthentication class.

This is a pure file/namespace reorganization: no types are renamed and
no behavior changes, so callers only need their using directives
updated.

Copilot-Session: 5b041bf4-6167-4ad5-99c1-b622f5a5a5f2
